Move tests into own file
